requirements.txt
# AI API
anthropic>=0.39.0

# HTTP Client
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.2
